Only delete .env if it exists

+++ lib/tasks/foreman.rake
@@ -1,6 +1,7 @@
 desc "Generate default .env file"
 task ".env" do
-  File.unlink(".env")
+  File.unlink(".env") if File.exists?(".env")
+
   File.open(".env", "w") do |f|
     f.puts <<ENVFILE
 RACK_ENV=development
